Move image classifier to its own controller
- it can't be called on Windows App Service because GDI+ is blocked
